+++ blackdoc/formats/doctest.py
@@ -1,10 +1,14 @@
 import itertools
+import re
 
 import more_itertools
 
 name = "doctest"
-prompt = ">>> "
-continuation_prompt = "... "
+prompt_length = 4
+prompt = ">>>"
+prompt_re = re.compile(r"(>>> ?)")
+continuation_prompt = "..."
+continuation_prompt_re = re.compile(r"(\.\.\. ?)")
 include_pattern = r"\.pyi?$"
 
 
@@ -16,7 +20,7 @@ def continuation_lines(lines):
             line_number = -1
             line = ""
 
-        if not line.lstrip().startswith(continuation_prompt):
+        if not continuation_prompt_re.match(line.lstrip()):
             break
 
         # actually consume the item
@@ -31,7 +35,7 @@ def detection_func(lines):
     except StopIteration:
         line = ""
 
-    if not line.lstrip().startswith(prompt):
+    if not prompt_re.match(line.lstrip()):
         return None
 
     detected_lines = list(
@@ -46,23 +50,71 @@ def detection_func(lines):
     return line_range, name, "\n".join(lines)
 
 
+def detect_docstring_quotes(line):
+    if "'''" in line:
+        docstring_quotes = "'''"
+    elif '"""' in line:
+        docstring_quotes = '"""'
+    else:
+        docstring_quotes = None
+
+    return docstring_quotes
+
+
 def extraction_func(line):
+    def extract_prompt(line):
+        match = prompt_re.match(line)
+        if match is not None:
+            (prompt,) = match.groups()
+            return prompt
+
+        match = continuation_prompt_re.match(line)
+        if match is not None:
+            (prompt,) = match.groups()
+            return prompt
+
+        return ""
+
+    def remove_prompt(line):
+        prompt = extract_prompt(line)
+        return line[len(prompt) :]
+
     lines = line.split("\n")
-    if any(line[:4] not in (prompt, continuation_prompt) for line in lines):
+    if any(
+        extract_prompt(line).rstrip() not in (prompt, continuation_prompt)
+        for line in lines
+    ):
         raise RuntimeError(f"misformatted code unit: {line}")
 
-    extracted_line = "\n".join(line[4:] for line in lines)
+    extracted_line = "\n".join(remove_prompt(line) for line in lines)
+    docstring_quotes = detect_docstring_quotes(extracted_line)
 
-    return {"prompt_length": len(prompt)}, extracted_line
+    return {
+        "prompt_length": len(prompt) + 1,
+        "docstring_quotes": docstring_quotes,
+    }, extracted_line
 
 
-def reformatting_func(line):
+def reformatting_func(line, docstring_quotes):
+    def add_prompt(prompt, line):
+        if not line:
+            return prompt
+
+        return " ".join([prompt, line])
+
     lines = iter(line.split("\n"))
 
     reformatted = "\n".join(
         itertools.chain(
-            more_itertools.always_iterable(prompt + more_itertools.first(lines)),
-            (continuation_prompt + line for line in lines),
+            more_itertools.always_iterable(
+                add_prompt(prompt, more_itertools.first(lines))
+            ),
+            (add_prompt(continuation_prompt, line) for line in lines),
         )
     )
+    # make sure nested docstrings still work
+    current_quotes = detect_docstring_quotes(reformatted)
+    if docstring_quotes != current_quotes:
+        reformatted = reformatted.replace(current_quotes, docstring_quotes)
+
     return reformatted