+++ blackdoc/tests/test_doctest.py
@@ -9,6 +9,19 @@ from .data.doctest import lines
 
 
 @pytest.mark.parametrize(
+    ("string", "expected"),
+    (
+        pytest.param("a", None, id="no quotes"),
+        pytest.param("'''a'''", "'''", id="single quotes"),
+        pytest.param('"""a"""', '"""', id="double quotes"),
+    ),
+)
+def test_detect_docstring_quotes(string, expected):
+    actual = doctest.detect_docstring_quotes(string)
+    assert expected == actual
+
+
+@pytest.mark.parametrize(
     "lines,expected",
     (
         pytest.param(lines[0], None, id="no_line"),
@@ -17,6 +30,14 @@ from .data.doctest import lines
             lines[4:8],
             ((1, 5), doctest.name, "\n".join(lines[4:8])),
             id="multiple_lines",
+        ),
+        pytest.param(
+            lines[23], ((1, 2), doctest.name, lines[23]), id="single empty line"
+        ),
+        pytest.param(
+            lines[17:23],
+            ((1, 7), doctest.name, "\n".join(lines[17:23])),
+            id="multiple lines with empty continuation line",
         ),
     ),
 )
@@ -32,14 +53,20 @@ def test_detection_func(lines, expected):
 @pytest.mark.parametrize(
     "line",
     (
-        pytest.param(textwrap.dedent(lines[8]), id="single_line"),
-        pytest.param(textwrap.dedent("\n".join(lines[4:8])), id="multiple_lines"),
+        pytest.param(textwrap.dedent(lines[8]), id="single line"),
+        pytest.param(textwrap.dedent(lines[23]), id="single empty line"),
+        pytest.param(textwrap.dedent("\n".join(lines[4:8])), id="multiple lines"),
+        pytest.param(
+            textwrap.dedent("\n".join(lines[17:23])),
+            id="multiple lines with empty continuation line",
+        ),
     ),
 )
 def test_extraction_func(line):
-    prompt_length = len(doctest.prompt)
+    docstring_quotes = doctest.detect_docstring_quotes(line)
+
     expected = (
-        {"prompt_length": prompt_length},
+        {"prompt_length": doctest.prompt_length, "docstring_quotes": docstring_quotes},
         "\n".join(line.lstrip()[4:] for line in line.split("\n")),
     )
     actual = doctest.extraction_func(line)
@@ -50,12 +77,25 @@ def test_extraction_func(line):
 @pytest.mark.parametrize(
     "expected",
     (
-        pytest.param(textwrap.dedent(lines[8]), id="single_line"),
-        pytest.param(textwrap.dedent("\n".join(lines[4:8])), id="multiple_lines"),
+        pytest.param(textwrap.dedent(lines[8]), id="single line"),
+        pytest.param(textwrap.dedent(lines[23]), id="single empty line"),
+        pytest.param(textwrap.dedent("\n".join(lines[4:8])), id="multiple lines"),
+        pytest.param(
+            textwrap.dedent("\n".join(lines[17:23])),
+            id="multiple lines with empty continuation line",
+        ),
+        pytest.param(
+            textwrap.dedent("\n".join(lines[17:23]).replace("'''", '"""')),
+            id="multiple lines with inverted docstring quotes",
+        ),
     ),
 )
 def test_reformatting_func(expected):
+    docstring_quotes = doctest.detect_docstring_quotes(expected)
     line = "\n".join(line.lstrip()[4:] for line in expected.split("\n"))
 
-    actual = doctest.reformatting_func(line)
+    actual = doctest.reformatting_func(line, docstring_quotes)
     assert expected == actual
+
+    # make sure the docstring quotes were not changed
+    assert docstring_quotes is None or docstring_quotes in actual