+++ doc/changelog.rst
@@ -3,7 +3,10 @@ Changelog
 
 v0.2 (*unreleased*)
 -------------------
-- Support the :rst:dir:`testcode`, :rst:dir:`testsetup` and :rst:dir:`testcleanup` directives.
+- Support the :rst:dir:`testcode`, :rst:dir:`testsetup` and
+  :rst:dir:`testcleanup` directives (:pull:`39`).
+- fix working with lines containing only the prompt and avoid changing the
+  quotes of nested docstrings (:issue:`41`, :pull:`43`)
 
 
 v0.1.2 (31 August 2020)